forgot geomfromwkb
